updated auth and frontend

+++ backend/controllers/UserController.js
@@ -1,60 +1,94 @@
 import User from "../models/UserModel.js";
+import argon2 from "argon2";
 
 export const getUsers = async(req, res) => {
     try {
-        const response = await User.findAll()
-        res.status(200).json(response)
+        const response = await User.findAll({
+          attributes: ['uuid', 'username', 'role']
+        });
+        res.status(200).json(response);
     } catch (error) {
-        console.log(error.message)
+        res.status(500).json({msg: error.message});
     }
 }
 
 export const getUsersById = async(req, res) => {
     try {
         const response = await User.findOne({
+            attributes: ['uuid', 'username', 'role'],
             where: {
-                id: req.params.id
+                uuid: req.params.id
             }
-        })
+        });
         res.status(200).json(response)
     } catch (error) {
-        console.log(error.message)
+      res.status(500).json({msg: error.message});
     }
 }
 
 export const createUser = async (req, res) => {
+    const {username, password, confirmPassword, role} = req.body;
+    if(password !== confirmPassword) return res.status(400).json({msg: "Password dan Confirm Password Tidak Cocok"});
+
+    const hashPassword = await argon2.hash(password);
     try {
-      await User.create(req.body);
-      res.status(201).json({msg: "User Created"});
+      await User.create({
+        username: username,
+        password: hashPassword,
+        role: role
+      });
+      res.status(201).json({msg: "Register Berhasil"});
     } catch (error) {
-      console.log(error.message);
+      res.status(400).json({msg: error.message});
     }
-  };
+  }
   
   export const updateUser = async (req, res) => {
+    const user = await User.findOne({
+      where: {
+          uuid: req.params.id
+      }
+    });
+    if(!user) return res.status(404).json({msg: "User tidak ditemukan"});
+    const {username, password, confirmPassword, role} = req.body;
+    let hashPassword;
+    if(password === "" || password === null) {
+      hashPassword = user.password
+    } else {
+      hashPassword = await argon2.hash(password);
+    }
+    if(password !== confirmPassword) return res.status(400).json({msg: "Password dan Confirm Password Tidak Cocok"});
     try {
-      await User.update(req.body, {
+      await User.update({
+        username: username,
+        password: hashPassword,
+        role: role
+      },{
         where: {
-          id: req.params.id
+          id: user.id
         }
       });
-      res.status(200).json({ msg: "User Updated" });
+      res.status(200).json({msg: "User Updated"});
     } catch (error) {
-      console.log(error.message);
-      res.status(500).json({ error: "Internal server error" });
+      res.status(400).json({msg: error.message});
     }
   };
   
   export const deleteUser = async (req, res) => {
+    const user = await User.findOne({
+      where: {
+          uuid: req.params.id
+      }
+    });
+    if(!user) return res.status(404).json({msg: "User tidak ditemukan"});
     try {
       await User.destroy({
         where: {
-          id: req.params.id
+          id: user.id
         }
       });
-      res.status(200).json({ msg: "User Deleted" });
+      res.status(200).json({msg: "User Deleted"});
     } catch (error) {
-      console.log(error.message);
-      res.status(500).json({ error: "Internal server error" });
+      res.status(400).json({msg: error.message});
     }
   };
